Split the plugin into various autoplugins.
This is a first step towards #30, but still very much a WIP.
