+++ src/main/scala/sbtunidoc/GenJavaUnidocKeys.scala
@@ -1,0 +1,7 @@
+package sbtunidoc
+
+import sbt.settingKey
+
+trait GenJavaUnidocKeys {
+  val unidocGenjavadocVersion = settingKey[String]("Version of the genjavadoc compiler plugin.")
+}
